requirements.txt
google-genai
faiss-cpu
numpy
pydantic
sqlalchemy
openai
